requirement.txt
backports.csv==1.0.7
certifi==2019.6.16
chardet==3.0.4
defusedxml==0.6.0
diff-match-patch==20181111
Django==2.0
django-ckeditor==5.4.0
django-crispy-forms==1.7.2
django-formtools==2.1
django-import-export==1.2.0
django-js-asset==1.2.2
django-mdeditor==0.1.14
django-ranged-response==0.2.0
django-reversion==3.0.4
django-simple-captcha==0.5.11
et-xmlfile==1.0.1
future==0.15.2
httplib2==0.9.2
idna==2.8
jdcal==1.4.1
Markdown==3.1.1
mysqlclient==1.4.2.post1
odfpy==1.4.0
openpyxl==2.6.2
Pillow==5.0.0
Pygments==2.4.2
pytz==2018.9
PyYAML==5.1.1
qiniu==7.2.4
requests==2.22.0
six==1.10.0
tablib==0.13.0
urllib3==1.25.3
xlrd==1.2.0
xlwt==1.3.0
